move claim integration tests out of keeper
