+++ x/incentive/integration_test.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/kava-labs/kava/app"
 	"github.com/kava-labs/kava/x/cdp"
+	"github.com/kava-labs/kava/x/incentive/testutil"
 	"github.com/kava-labs/kava/x/pricefeed"
 )
 
@@ -139,80 +140,72 @@ func NewPricefeedGenStateMulti() app.GenesisState {
 	return app.GenesisState{pricefeed.ModuleName: pricefeed.ModuleCdc.MustMarshalJSON(pfGenesis)}
 }
 
-// func NewIncentiveGenState(previousAccumTime, endTime time.Time, rewardPeriods ...incentive.RewardPeriod) app.GenesisState {
-// 	var accumulationTimes incentive.GenesisAccumulationTimes
-// 	for _, rp := range rewardPeriods {
-// 		accumulationTimes = append(
-// 			accumulationTimes,
-// 			incentive.NewGenesisAccumulationTime(
-// 				rp.CollateralType,
-// 				previousAccumTime,
-// 			),
-// 		)
-// 	}
-// 	genesis := incentive.NewGenesisState(
-// 		incentive.NewParams(
-// 			rewardPeriods,
-// 			types.MultiRewardPeriods{},
-// 			types.MultiRewardPeriods{},
-// 			types.MultiRewardPeriods{},
-// 			types.MultiRewardPeriods{},
-// 			incentive.Multipliers{
-// 				incentive.NewMultiplier(incentive.Small, 1, d("0.25")),
-// 				incentive.NewMultiplier(incentive.Large, 12, d("1.0")),
-// 			},
-// 			endTime,
-// 		),
-// 		accumulationTimes,
-// 		accumulationTimes,
-// 		accumulationTimes,
-// 		incentive.DefaultGenesisAccumulationTimes,
-// 		incentive.DefaultUSDXClaims,
-// 		incentive.DefaultHardClaims,
-// 	)
-// 	return app.GenesisState{incentive.ModuleName: incentive.ModuleCdc.MustMarshalJSON(genesis)}
-// }
-
-func NewCDPGenStateHighInterest() app.GenesisState {
-	cdpGenesis := cdp.GenesisState{
-		Params: cdp.Params{
-			GlobalDebtLimit:         sdk.NewInt64Coin("usdx", 2000000000000),
-			SurplusAuctionThreshold: cdp.DefaultSurplusThreshold,
-			SurplusAuctionLot:       cdp.DefaultSurplusLot,
-			DebtAuctionThreshold:    cdp.DefaultDebtThreshold,
-			DebtAuctionLot:          cdp.DefaultDebtLot,
-			CollateralParams: cdp.CollateralParams{
-				{
-					Denom:               "bnb",
-					Type:                "bnb-a",
-					LiquidationRatio:    sdk.MustNewDecFromStr("1.5"),
-					DebtLimit:           sdk.NewInt64Coin("usdx", 500000000000),
-					StabilityFee:        sdk.MustNewDecFromStr("1.000000051034942716"), // 500% APR
-					LiquidationPenalty:  d("0.05"),
-					AuctionSize:         i(50000000000),
-					Prefix:              0x22,
-					SpotMarketID:        "bnb:usd",
-					LiquidationMarketID: "bnb:usd",
-					ConversionFactor:    i(8),
-				},
-			},
-			DebtParam: cdp.DebtParam{
-				Denom:            "usdx",
-				ReferenceAsset:   "usd",
-				ConversionFactor: i(6),
-				DebtFloor:        i(10000000),
+func NewPricefeedGenStateMultiFromTime(t time.Time) app.GenesisState {
+	pfGenesis := pricefeed.GenesisState{
+		Params: pricefeed.Params{
+			Markets: []pricefeed.Market{
+				{MarketID: "kava:usd", BaseAsset: "kava", QuoteAsset: "usd", Oracles: []sdk.AccAddress{}, Active: true},
+				{MarketID: "btc:usd", BaseAsset: "btc", QuoteAsset: "usd", Oracles: []sdk.AccAddress{}, Active: true},
+				{MarketID: "xrp:usd", BaseAsset: "xrp", QuoteAsset: "usd", Oracles: []sdk.AccAddress{}, Active: true},
+				{MarketID: "bnb:usd", BaseAsset: "bnb", QuoteAsset: "usd", Oracles: []sdk.AccAddress{}, Active: true},
+				{MarketID: "busd:usd", BaseAsset: "busd", QuoteAsset: "usd", Oracles: []sdk.AccAddress{}, Active: true},
+				{MarketID: "zzz:usd", BaseAsset: "zzz", QuoteAsset: "usd", Oracles: []sdk.AccAddress{}, Active: true},
 			},
 		},
-		StartingCdpID: cdp.DefaultCdpStartingID,
-		DebtDenom:     cdp.DefaultDebtDenom,
-		GovDenom:      cdp.DefaultGovDenom,
-		CDPs:          cdp.CDPs{},
-		PreviousAccumulationTimes: cdp.GenesisAccumulationTimes{
-			cdp.NewGenesisAccumulationTime("bnb-a", time.Time{}, sdk.OneDec()),
-		},
-		TotalPrincipals: cdp.GenesisTotalPrincipals{
-			cdp.NewGenesisTotalPrincipal("bnb-a", sdk.ZeroInt()),
+		PostedPrices: []pricefeed.PostedPrice{
+			{
+				MarketID:      "kava:usd",
+				OracleAddress: sdk.AccAddress{},
+				Price:         sdk.MustNewDecFromStr("2.00"),
+				Expiry:        t.Add(1 * time.Hour),
+			},
+			{
+				MarketID:      "btc:usd",
+				OracleAddress: sdk.AccAddress{},
+				Price:         sdk.MustNewDecFromStr("8000.00"),
+				Expiry:        t.Add(1 * time.Hour),
+			},
+			{
+				MarketID:      "xrp:usd",
+				OracleAddress: sdk.AccAddress{},
+				Price:         sdk.MustNewDecFromStr("0.25"),
+				Expiry:        t.Add(1 * time.Hour),
+			},
+			{
+				MarketID:      "bnb:usd",
+				OracleAddress: sdk.AccAddress{},
+				Price:         sdk.MustNewDecFromStr("17.25"),
+				Expiry:        t.Add(1 * time.Hour),
+			},
+			{
+				MarketID:      "busd:usd",
+				OracleAddress: sdk.AccAddress{},
+				Price:         sdk.OneDec(),
+				Expiry:        t.Add(1 * time.Hour),
+			},
+			{
+				MarketID:      "zzz:usd",
+				OracleAddress: sdk.AccAddress{},
+				Price:         sdk.MustNewDecFromStr("2.00"),
+				Expiry:        t.Add(1 * time.Hour),
+			},
 		},
 	}
-	return app.GenesisState{cdp.ModuleName: cdp.ModuleCdc.MustMarshalJSON(cdpGenesis)}
+	return app.GenesisState{pricefeed.ModuleName: pricefeed.ModuleCdc.MustMarshalJSON(pfGenesis)}
+}
+
+func NewHardGenStateMulti(genTime time.Time) testutil.HardGenesisBuilder {
+	kavaMM := testutil.NewStandardMoneyMarket("ukava")
+	kavaMM.SpotMarketID = "kava:usd"
+	btcMM := testutil.NewStandardMoneyMarket("btcb")
+	btcMM.SpotMarketID = "btc:usd"
+
+	builder := testutil.NewHardGenesisBuilder().WithGenesisTime(genTime).
+		WithInitializedMoneyMarket(testutil.NewStandardMoneyMarket("usdx")).
+		WithInitializedMoneyMarket(kavaMM).
+		WithInitializedMoneyMarket(testutil.NewStandardMoneyMarket("bnb")).
+		WithInitializedMoneyMarket(btcMM).
+		WithInitializedMoneyMarket(testutil.NewStandardMoneyMarket("xrp")).
+		WithInitializedMoneyMarket(testutil.NewStandardMoneyMarket("zzz"))
+	return builder
 }
